整理 Id

+++ LockstepClient/Assets/LockstepCore/Game/Lockstep.Game.Features.Input/ExecuteSpawnInput.cs
@@ -49,7 +49,7 @@ namespace Lockstep.Game.Features.Input
 
 
                 //UnityEngine.Debug.Log("[ExecuteSpawnInput]    " + entityWithId.id.value + " -> " + value);
-                gameEntity.AddId(value);
+                gameEntity.AddId((byte)value);
                 gameEntity.AddActorId(item.actorId.value);
                 gameEntity.AddLocalId(_localIdCounter);
                 gameEntity.AddVelocity(Vector2.Zero);
